\name{misc}
\alias{misc}
\alias{RT}
\alias{TX}
\alias{bounded_outward_trajectory}
\alias{unbounded_outward_trajectory}
\alias{cone}
\alias{lightcone}
\alias{ingoing}
\alias{ingoing_light}
\alias{ingoing_null_arrow}
\alias{ingoing_trajectory}
\alias{inward_trajectory}
\alias{null_geodesics}
\alias{outgoing}
\alias{outgoing_null_arrow}
\alias{outward_inside_trajectory}
\alias{penrose_transform}
\alias{raindrop}
\alias{raindrop_arrow}
\alias{thrower}
\alias{thrower_asp1}
\alias{trajectory}
\title{
Miscellaneous helper functions
}
\description{
Various helper functions that should not appear in the plotting routines
}
\usage{
RT(TX)
TX(rt,exterior)
}
\arguments{
  \item{TX,rt}{Two-column matrix giving time and space}
  \item{exterior}{Boolean, specifying whether the point in question is
    inside or outside the event horizon}
}
\details{
details here
}
\author{Robin K. S. Hankin}
