\name{schwarzschild-package}
\alias{schwarzschild-package}
\docType{package}
\title{Diagrams of various spacetimes including black holes}
\description{
A collection of routines to draw mathematically correct diagrams of the
spacetime near a black hole; also some visualization of gravitational
waves
}
\details{
Most of the package functionality is the definition of functions such as
\code{schwarzschild()} which create mathematically precise diagrams of
black holes.
}
\author{Robin K. S. Hankin}
\keyword{ package }
\examples{
schwarzschild()
}
