\name{schwarzschild-package}
\alias{schwarzschild-package}
\docType{package}
\title{adslkdfja;lkjdf}
\description{
dsdffddfs}
\details{
  An overview of how to use the package, including the most important
  functions
}
\author{
dsfsdfdfsdsf
Maintainer: ssfdsdffds
}
\references{
Literature or other references for background information
}
\keyword{ package }
\examples{
schwarzschild()
}
