\name{schwarzschild-package}
\alias{schwarzschild-package}
\docType{package}
\title{
\packageTitle{schwarzschild}
}
\description{
\packageDescription{schwarzschild}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{schwarzschild}

Most of the package functionality is the definition of functions such as
\code{schwarzschild()} which produce spacetime diagrams.  The functions
are intended to be called by\code{inst/maker.R} which produces annotated
and correctly-sized PDF files.

\packageIndices{schwarzschild}
}
\author{
\packageAuthor{schwarzschild}

Maintainer: \packageMaintainer{schwarzschild}
}
\keyword{ package }
\examples{
schwarzschild()
}
