\name{schwarzschild-package}
\alias{schwarzschild-package}
\docType{package}
\title{
\packageTitle{schwarzschild}
}
\description{
\packageDescription{schwarzschild}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{schwarzschild}
\packageIndices{schwarzschild}
}
\author{
\packageAuthor{schwarzschild}

Maintainer: \packageMaintainer{schwarzschild}
}
\keyword{ package }
\examples{
schwarzschild()
}
